requirements.txt
python-dotenv
aiogram
fastapi
uvicorn
apscheduler
